Move level 0 scenario to test package
